# openapi.yaml
openapi: 3.0.3
info:
  title: Goods In — Inventory Receiving API
  description: >
    Scanner-first IMEI receiving, manifest reconciliation, SKU confirmation,
    device lifecycle management, and DYMO label printing for wholesale
    device traders, refurbishers, and graders.


    Every route under `/api/*` requires a valid JWT **except**
    `GET /api/health` and `POST /api/auth/dev-login`. All data is scoped to
    the caller's `organisation_id`, resolved server-side from the verified
    token — it is never accepted as a request parameter.


    Out of scope for this version of the API: grading workflow, OPR
    export/import documentation, C&E1154, commercial-invoice generation,
    and the export/return/sale lifecycle transitions (the corresponding
    device statuses exist in the enum but have no transitions wired to
    them yet).
  version: "1.1.0"
  license:
    name: Proprietary
servers:
  - url: /api
    description: Same-origin (the SPA and API are served from one Cloudflare Pages project)
security:
  - bearerAuth: []
...
paths:
...
  /scan:
    post:
      tags: [Scan]
      summary: Scan an IMEI against the active manifest
      description: >
        Server re-validates the IMEI (14-16 digits + Luhn at 15 digits)
        regardless of client-side checks. Returns one of four outcomes:
        `matched`, `duplicate`, `unreconciled`, or `rejected` (malformed
        IMEI).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest_id, imei]
              properties:
                manifest_id: { type: integer }
                imei: { type: string }
      responses:
        "200":
          description: Scan outcome
          content:
            application/json:
              schema:
                type: object
                properties:
                  outcome:
                    type: string
                    enum: [matched, duplicate, unreconciled, rejected]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /scan/confirm:
    post:
      tags: [Scan]
      summary: Confirm a matched SKU and receive the device
      description: >
        Valuation fields (`buy_price`, `currency`, `vat_type`) are
        **required** and validated authoritatively server-side
        (`src/lib/validate.ts`) — an unrecognised ISO 4217 currency code
        (e.g. `UKL`) is rejected with `422` regardless of what any client
        allowed through. On success, writes `received_devices`
        (`status=RECEIVED`), a `RECEIVE` `device_events` row, a
        `print_jobs` row, and marks the manifest line `received`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_device_id, sku, buy_price, currency, vat_type]
              properties:
                expected_device_id: { type: integer }
                sku: { type: string }
                brand: { type: string }
                model: { type: string }
                capacity: { type: string }
                color: { type: string }
                grade: { type: string, description: "A+, A, B+, B, C+, C, D, or UG" }
                notes: { type: string, nullable: true }
                auto_print: { type: boolean }
                buy_price: { type: number, minimum: 0, maximum: 1000000 }
                currency: { type: string, description: "ISO 4217, e.g. GBP", default: GBP }
                vat_type: { type: string, enum: [MARGIN, STANDARD, ZERO] }
                supplier_id: { type: integer, nullable: true }
      responses:
        "200":
          description: Device received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { $ref: "#/components/schemas/Device" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Missing/invalid valuation fields (e.g. unrecognised currency code)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /scan/force-add:
    post:
      tags: [Scan]
      summary: Force-add an off-manifest (unreconciled) IMEI to inventory
      description: >
        Same body shape as `/scan/confirm`, and the SAME valuation
        requirements: `buy_price` and `vat_type` are required, `currency`
        must be a valid ISO 4217 code (default GBP). The off-manifest
        exception branch cannot be used to bypass required valuation
        fields — a request missing them is rejected with 422 and no
        device/audit/print rows are written.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanConfirmRequest"
      responses:
        "200":
          description: Device force-added
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { $ref: "#/components/schemas/Device" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Missing or invalid valuation fields (buy_price / vat_type required; currency must be valid ISO 4217)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /scan/manual:
    post:
      tags: [Scan]
      summary: Manually add a device outside the scan flow
      description: >
        Same valuation requirements as `/scan/confirm` and `/scan/force-add`:
        `buy_price` and `vat_type` are required, `currency` must be a valid
        ISO 4217 code (default GBP). Every path that creates a device
        enforces these rules server-side — no intake branch is a bypass.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanConfirmRequest"
      responses:
        "200":
          description: Device added
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { $ref: "#/components/schemas/Device" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Missing or invalid valuation fields (buy_price / vat_type required; currency must be valid ISO 4217)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /scan/reject:
    post:
      tags: [Scan]
      summary: Audit-log a rejection (writes a REJECT device_events row)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                manifest_id: { type: integer }
                imei: { type: string }
                notes: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean } } }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /scan/events/{manifestId}:
    get:
      tags: [Scan]
      summary: Recent scan events for a manifest
      parameters:
        - name: manifestId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { type: object } }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256 JWT obtained from `POST /auth/dev-login`. Sent as
        `Authorization: Bearer <token>`. A subset of GET routes that are
        opened via `window.open()` (DYMO label pages) also accept the
        token as a `?token=` query parameter as a fallback.
...
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or malformed-claims token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    DeviceStatus:
      type: string
      enum:
        - RECEIVED
        - SORTING
        - ACTIVE_INVENTORY
        - IN_HOUSE_REPAIR
        - READY_FOR_EXPORT
        - IN_EXPORT_CONSIGNMENT
        - EXPORTED_UNDER_OPR
        - RETURNED_UNDER_OPR
        - SOLD
        - REJECTED
      description: >
        Allowed transitions (this pass only): RECEIVED→SORTING,
        SORTING→{ACTIVE_INVENTORY,IN_HOUSE_REPAIR,READY_FOR_EXPORT},
        IN_HOUSE_REPAIR→ACTIVE_INVENTORY, RECEIVED→REJECTED. All other
        statuses/transitions exist in the enum for forward-compatibility
        but have no transitions wired to them yet (out of scope).
...
    Device:
      type: object
      description: A row from received_devices.
      properties:
        id: { type: integer }
        organisation_id: { type: integer }
        uuid: { type: string, description: 12-char short code }
        imei: { type: string }
        sku: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        capacity: { type: string, nullable: true }
        color: { type: string, nullable: true }
        grade: { type: string, nullable: true }
        source: { type: string, enum: [manifest, unreconciled] }
        manifest_id: { type: integer, nullable: true }
        expected_device_id: { type: integer, nullable: true }
        status: { $ref: "#/components/schemas/DeviceStatus" }
        label_printed_at: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
        created_by_user_id: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        buy_price: { type: number, nullable: true }
        currency: { type: string, example: GBP }
        vat_type: { type: string, enum: [MARGIN, STANDARD, ZERO], nullable: true }
        supplier_id: { type: integer, nullable: true }
...
    ScanConfirmRequest:
      type: object
      properties:
        expected_device_id: { type: integer, nullable: true }
        sku: { type: string }
        brand: { type: string }
        model: { type: string }
        capacity: { type: string }
        color: { type: string }
        grade: { type: string }
        notes: { type: string, nullable: true }
        auto_print: { type: boolean }
        buy_price: { type: number, nullable: true }
        currency: { type: string, nullable: true, default: GBP }
        vat_type: { type: string, enum: [MARGIN, STANDARD, ZERO], nullable: true }
        supplier_id: { type: integer, nullable: true }
...